replace uint on int in Pos struct

+++ example/example.go
@@ -23,7 +23,7 @@ func (g greg) Start(anthill pkg.Pos, birth pkg.Pos) {
 }
 
 func (g greg) Do(fields [5][5]pkg.FieldType, round int) (target pkg.Pos, action pkg.Action) {
-	target = pkg.Pos{uint(r.Intn(5)), uint(r.Intn(5))}
+	target = pkg.Pos{r.Intn(5), r.Intn(5)}
 	action = pkg.ResolveAction(fields[target.X()][target.Y()])
 
 	return
